+++ src/app/api/auth/[...nextauth]/route.js
@@ -1,8 +1,4 @@
-import getAuthOptions from "@/app/authOptions";
-import NextAuth from "next-auth";
-
-async function handler(req, res) {
-  return NextAuth(req, res, await getAuthOptions());
-}
-
-export { handler as GET, handler as POST };
+export {
+  authHandler as GET,
+  authHandler as POST,
+} from "@/backend/routes/auth/authHandler";
