+++ src/app/api/items/[id]/route.js
@@ -1,8 +1,1 @@
-import { NextResponse } from "next/server";
-import getItemsController from "@/backend/entities/items";
-
-export async function GET(req, { params: { id } }) {
-  const getItemById = await getItemsController("GET_ONE_ITEM");
-  const res = await getItemById(id);
-  return NextResponse.json(res, { status: res.status });
-}
+export { getItemById as GET } from "@/backend/routes/items/getItemById";
