+++ src/backend/db/connectDB.js
@@ -3,6 +3,7 @@ import mongoose from "mongoose";
 const { HOST_DB } = process.env;
 
 export default async function connectDB() {
+  console.log(mongoose.connection.readyState);
   if (mongoose.connection.readyState === 1) {
     return;
   }
@@ -14,3 +15,5 @@ export default async function connectDB() {
     console.log("database error connection:", error);
   }
 }
+
+connectDB();
