+++ src/backend/entities/items/helpers/getTryCatchWrapper.js
@@ -1,6 +1,6 @@
 import { getNotFoundItemError } from "./getErrorObj";
 
-export async function getTryCatchWrapper(callback) {
+export function getTryCatchWrapper(callback) {
   return async function (...args) {
     try {
       const res = await callback(...args);
