+++ src/backend/entities/orders/helpers/getTryCatchWrapper.js
@@ -1,6 +1,6 @@
 import { getOrderError } from "./getErrorObj";
 
-export async function getTryCatchWrapper(callback) {
+export function getTryCatchWrapper(callback) {
   return async function (...args) {
     try {
       const res = await callback(...args);
