+++ src/backend/entities/orders/index.js
@@ -1,14 +1,13 @@
-import { getOrderControllers } from "./getOrderControllers";
+import { orderControllers } from "./orderControllers";
 import { getTryCatchWrapper } from "./helpers";
 
-export default async function getOrderController(method) {
-  const orderControllers = await getOrderControllers();
+export default function getOrderController(method) {
   switch (method) {
     case "GET_ORDER_BY_FIELD":
-      return await getTryCatchWrapper(orderControllers.getOrderByField);
+      return getTryCatchWrapper(orderControllers.getOrderByField);
     case "GET_ORDER_BY_ID":
-      return await getTryCatchWrapper(orderControllers.getOrderById);
+      return getTryCatchWrapper(orderControllers.getOrderById);
     case "CREATE_ORDER":
-      return await getTryCatchWrapper(orderControllers.createOrder);
+      return getTryCatchWrapper(orderControllers.createOrder);
   }
 }
