+++ src/backend/routes/auth/authHandler.js
@@ -1,0 +1,6 @@
+import getAuthOptions from "@/backend/libs/authOptions";
+import NextAuth from "next-auth";
+
+export async function authHandler(req, res) {
+  return NextAuth(req, res, getAuthOptions());
+}
